# docs/swagger.yaml
basePath: /
definitions:
...
  appconfig.LoggingConfig:
    properties:
      endpoint:
        type: string
      server_level:
        type: string
      sink:
        type: string
      stream:
        type: string
    type: object
...
  handlers.AcceptedResponse:
    properties:
      correlation_id:
        type: string
      event:
        type: string
      status:
        type: string
    type: object
...
  handlers.TaskRequest:
    properties:
      command:
        type: string
    type: object
...
  handlers.setAgentLogLevelRequest:
    properties:
      log_level:
        type: string
    type: object
  logging.Record:
    properties:
      attrs:
        additionalProperties: {}
        type: object
      level:
        type: string
      message:
        type: string
      source:
        type: string
      time:
        type: string
    type: object
...
info:
  contact: {}
  description: Event-driven Go API backed by MongoDB and Redis. Every request carries
    a correlation ID (X-Correlation-ID), generated by the server if the caller doesn't
    supply one.
  title: Metarr API
  version: "1.0"
paths:
...
  /api/config/agents/{slug}/log-level:
    post:
      consumes:
      - application/json
      description: Switches a single agent between info and debug. If the agent isn't
        configured with any libraries yet, this creates a minimal entry for it rather
        than requiring it be configured first.
      parameters:
      - description: Agent slug
        in: path
        name: slug
        required: true
        type: string
      - description: Log level
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.setAgentLogLevelRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handlers.AcceptedResponse'
        "400":
          description: invalid request body or log_level
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Set one agent's log level
      tags:
      - Config
...
  /api/config/logging:
    get:
      description: Returns the server's own log level and the informational pipeline
        fields shown on the System > Logging screen. Endpoint/sink/stream are informational
        only — the actual pipeline is Fluent Bit, configured independently of this
        document.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/appconfig.LoggingConfig'
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Fetch the logging configuration
      tags:
      - Config
    post:
      consumes:
      - application/json
      description: 'Replaces the logging section: the server''s own log level, and
        the informational sink/endpoint/stream fields. Changing the informational
        fields does not reconfigure Fluent Bit — see the LoggingConfig doc comment.'
      parameters:
      - description: Logging configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/appconfig.LoggingConfig'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handlers.AcceptedResponse'
        "400":
          description: invalid request body or server_level
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Update the logging configuration
      tags:
      - Config
...
  /api/logging/tail:
    get:
      description: Returns the most recent log records seen on the centralized logging
        channel, from every process (server and every agent) that has published to
        it.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/logging.Record'
            type: array
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Fetch the recent log tail
      tags:
      - Config
...
  /api/tasks/directory-scan/{slug}:
    post:
      consumes:
      - application/json
      description: Scans the configured scan directory with the given scanner_slug,
        storing one record per media item directory plus one per media file in the
        local_directory collection. Fires the directory_scan event onto the durable
        event stream and returns as soon as it has been queued.
      parameters:
      - description: scanner_slug of the scan directory to scan
        in: path
        name: slug
        required: true
        type: string
      - description: Command to run
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TaskRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handlers.AcceptedResponse'
        "400":
          description: invalid request body, unsupported command, or the entry has
            an unusable scan_type
          schema:
            type: string
        "404":
          description: no scan directory with that slug
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Trigger a directory scan
      tags:
      - Tasks
  /api/tasks/sonarr_cache_data:
    post:
      consumes:
      - application/json
      description: Fires the sonarr_cache_data event onto the durable event stream
        in a non-blocking way and returns as soon as the event has been queued.
      parameters:
      - description: Command to run
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TaskRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handlers.AcceptedResponse'
        "400":
          description: invalid request body or unsupported command
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Trigger the sonarr_cache_data background job
      tags:
      - Tasks
...
securityDefinitions:
  ApiKeyHeaderAuth:
    description: API key via the X-Api-Key header.
    in: header
    name: X-Api-Key
    type: apiKey
  ApiKeyQueryAuth:
    description: API key via the apikey query parameter. Either this or ApiKeyHeaderAuth
      satisfies auth on protected routes.
    in: query
    name: apikey
    type: apiKey
swagger: "2.0"